refactor: apply review feedback to the web3 confirmation popup
Rename the popup entry points to ShowForResultAsync, fold the decoder
into DecodedTransaction as From/TryFromMetaTransaction behind an
internal constructor, take a non-null payload in ShowRawPayload, hold
the MANA contracts in a ReadOnlyHashSet, and lift the parcel ternary
into SceneNameFromParcelAsync.
