+++ Explorer/Assets/DCL/AuthenticationScreenFlow/TransactionRecipientUtils.cs
@@ -1,3 +1,4 @@
+using DCL.Utility;
 using DCL.Web3;
 using System;
 using System.Collections.Generic;
@@ -6,23 +7,17 @@ using System.Numerics;
 namespace DCL.AuthenticationScreenFlow
 {
     /// <summary>
-    ///     Builds the plain-language copy of the recipient confirmation step: how much is being sent and
-    ///     to whom.
+    ///     Builds the plain-language copy of the recipient confirmation step.
     /// </summary>
     public static class TransactionRecipientUtils
     {
-        /// <summary>
-        ///     Copy for a request that matches no mapped shape, so neither the amount nor the recipient is
-        ///     known and the raw payload of the first step is all there is to go on.
-        /// </summary>
         public const string UNKNOWN_REQUEST_DESCRIPTION =
             "Are you sure you want to approve this request? Decentraland cannot tell what it does or who receives your assets.";
 
-        // Link/highlight blue used for names and addresses in the confirmation copy.
         private const string HIGHLIGHT_COLOR = "#32CEFF";
 
-        // MANA renders as a sprite. The name is matched case-sensitively against the character table of
-        // the description's sprite asset (PolygonManaIcon), which spells it "Mana"; anything else is tofu.
+        // TMP matches sprite names case-sensitively against the PolygonManaIcon character table, which
+        // spells it "Mana"; anything else is tofu.
         private const string MANA_SPRITE = "<sprite name=\"Mana\">";
 
         // The popup labels gas and balance in "ETH"; keep the native transfer symbol consistent.
@@ -32,10 +27,9 @@ namespace DCL.AuthenticationScreenFlow
         private const int TOKEN_DECIMALS = 18;
         private const int MAX_FRACTION_DIGITS = 4;
 
-        // Known Decentraland MANA contracts (https://contracts.decentraland.org/addresses.json, same
-        // source as DonationsService). Any network matches: the symbol is display-only and a scene may
-        // transfer MANA on either network.
-        private static readonly HashSet<string> MANA_CONTRACTS = new (StringComparer.OrdinalIgnoreCase)
+        // https://contracts.decentraland.org/addresses.json, same source as DonationsService. Any network
+        // matches: the symbol is display-only and a scene may transfer MANA on either.
+        private static readonly ReadOnlyHashSet<string> MANA_CONTRACTS = new (new HashSet<string>(StringComparer.OrdinalIgnoreCase)
         {
             "0x0f5d2fb29fb7d3cfee444a200298f468908cc942", // Ethereum Mainnet
             "0xe7fdae84acaba2a5ba817b6e6d8a2d415dbfedbe", // Ethereum Goerli
@@ -43,7 +37,7 @@ namespace DCL.AuthenticationScreenFlow
             "0xa1c57f48f0deb89f569dfbe6e2b7f46d33606fd4", // Polygon Mainnet (PoS)
             "0x882da5967c435ea5cc6b09150d55e8304b838f45", // Polygon Mumbai Testnet
             "0x7ad72b9f944ea9793cf4055d88f81138cc2c63a0", // Polygon Amoy Testnet
-        };
+        });
 
         public static string ProfileDescription(string amount, string address, string name) =>
             $"Are you sure you want to send {amount} to {HighlightLink(address, "@" + name)}?";
@@ -55,8 +49,7 @@ namespace DCL.AuthenticationScreenFlow
             $"Are you sure you want to send {amount} to a wallet outside of Decentraland: {Highlight(address)}?";
 
         /// <summary>
-        ///     "5 &lt;MANA sprite&gt;" for MANA, "5 ETH" for the native currency, or "assets" when the
-        ///     amount cannot be determined (an unknown token or an opaque contract call).
+        ///     "5 &lt;MANA sprite&gt;", "5 ETH", or "assets" when the amount cannot be determined.
         /// </summary>
         public static string Amount(DecodedTransaction decoded)
         {
@@ -83,9 +76,8 @@ namespace DCL.AuthenticationScreenFlow
             BigInteger whole = amount / divisor;
             string fraction = (amount % divisor).ToString().PadLeft(TOKEN_DECIMALS, '0')[..MAX_FRACTION_DIGITS].TrimEnd('0');
 
-            // Anything under 0.0001 rounds away at this precision, and "0" on a confirmation reads as
-            // sending nothing at all. Worded rather than written "<0.0001": the amount is placed into
-            // rich text unescaped so it can carry the MANA sprite, so it must not introduce a '<'.
+            // "0" on a confirmation reads as sending nothing at all. Worded rather than written "<0.0001"
+            // because the amount is interpolated into rich text unescaped, so it must not introduce a '<'.
             if (whole.IsZero && fraction.Length == 0)
                 return "under 0.0001";
 
@@ -99,9 +91,8 @@ namespace DCL.AuthenticationScreenFlow
             $"<link=\"{EscapeRichText(id)}\"><color={HIGHLIGHT_COLOR}><b>{EscapeRichText(label)}</b></color></link>";
 
         /// <summary>
-        ///     Swaps the characters TMP reads as markup for lookalikes that it does not. The copy is
-        ///     rendered as rich text by design (the MANA sprite, the profile link), so a display name or
-        ///     scene name carrying "&lt;size=0&gt;" would otherwise hide the warning it appears inside.
+        ///     Swaps the characters TMP reads as markup for lookalikes that it does not. The copy is rich
+        ///     text by design, so a display name carrying "&lt;size=0&gt;" would hide the warning it sits in.
         /// </summary>
         private static string EscapeRichText(string value) =>
             value.Replace('<', '‹') // single left-pointing angle quotation mark
