+++ Explorer/Assets/DCL/Web3/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("DCL.EditMode.Tests")]
