+++ Explorer/Assets/DCL/Web3/AssemblyInfo.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2a9df9a8c802f264bb75a238bd2af8f5
